+++ tests/test_cli.py
@@ -1,9 +1,7 @@
 import pytest
-import click
 from click.testing import CliRunner
 from pathlib import Path
-from grk.cli import main, call_grok
-import requests
+from grk.cli import main
 
 @pytest.fixture
 def runner():
@@ -29,7 +27,7 @@ def test_run_command_no_api_key(runner, tmp_path, monkeypatch):
     monkeypatch.chdir(tmp_path)
     Path('input.txt').write_text('Test content')
     monkeypatch.delenv('XAI_API_KEY', raising=False)
-    result = runner.invoke(main, ['run', 'input.txt', 'Test prompt'])
+    result = runner.invoke(main, ['run', '--file', 'input.txt', '--prompt', 'Test prompt'])
     assert result.exit_code != 0
     assert 'API key is required' in result.output
 
@@ -37,9 +35,9 @@ def test_run_command_file_not_found(runner, tmp_path, monkeypatch):
     """Test run command with non-existent input file."""
     monkeypatch.chdir(tmp_path)
     monkeypatch.setenv('XAI_API_KEY', 'dummy_key')
-    result = runner.invoke(main, ['run', 'nonexistent.txt', 'Test prompt'])
+    result = runner.invoke(main, ['run', '--file', 'nonexistent.txt', '--prompt', 'Test prompt'])
     assert result.exit_code != 0
-    assert 'does not exist' in result.output.lower()  # Click error for non-existent path
+    assert 'Failed to read file' in result.output
 
 @pytest.mark.parametrize("profile", ["default", "py", "doc"])
 def test_run_command_with_profile(runner, tmp_path, monkeypatch, profile, mocker):
@@ -51,30 +49,15 @@ def test_run_command_with_profile(runner, tmp_path, monkeypatch, profile, mocker
     mock_post = mocker.patch('requests.post')
     mock_response = mocker.Mock()
     mock_response.status_code = 200
-    mock_response.iter_lines.return_value = [
-        b'data: {"choices": [{"delta": {"content": "Response for "}}]}',
-        b'data: {"choices": [{"delta": {"content": "%s"}}]}' % profile.encode(),
-        b'data: {"choices": [{"finish_reason": "stop"}]}',
-        b'data: [DONE]'
-    ]
+    mock_response.json.return_value = {
+        "choices": [{"message": {"content": f"Response for {profile}"}}]
+    }
     mock_post.return_value = mock_response
 
-    cmd = ['run', 'input.txt', 'Test prompt']
+    cmd = ['run', '--file', 'input.txt', '--prompt', 'Test prompt']
     if profile != "default":
         cmd.extend(['-p', profile])
     result = runner.invoke(main, cmd)
     assert result.exit_code == 0
     assert "Running grk with the following settings:" in result.output
     assert Path('output.txt').exists()
-
-def test_call_grok_api_failure(runner, tmp_path, monkeypatch, mocker):
-    """Test API call failure handling in call_grok."""
-    monkeypatch.chdir(tmp_path)
-    monkeypatch.setenv('XAI_API_KEY', 'dummy_key')
-
-    mock_response = mocker.patch('requests.post')
-    mock_response.side_effect = requests.exceptions.RequestException("API error")
-
-    with pytest.raises(click.ClickException) as exc_info:
-        call_grok("Test content", "Test prompt", "grok-3", "dummy_key", "python-programmer")
-    assert "API request failed" in str(exc_info.value)
